+++ docker/dockerfiles/DockerfileAdmin
@@ -10,8 +10,8 @@ RUN pip install --upgrade pip && pip install --no-cache-dir -r admin.txt
 COPY /standupuz_web .
 
 # Добавляем задержку перед запуском сервера
-CMD ["sh", "-c", "python manage.py runserver 0.0.0.0:8000"]
-#CMD ["sh", "-c", "python manage.py collectstatic && python manage.py runserver 0.0.0.0:8000"]
+#CMD ["sh", "-c", "python manage.py runserver 0.0.0.0:8000"]
+CMD ["sh", "-c", "python manage.py megrate && python manage.py runserver 0.0.0.0:8000"]
 
 #CMD ["python", "manage.py", "runserver", "0.0.0.0:8000"]
 #CMD ["sh", "-c", "python manage.py runserver 0.0.0.0:8000 --noreload"]
