+++ standupuz_web/events/utils.py
@@ -13,3 +13,9 @@ def get_photo_url(photo_id: str) -> str:
     return f'../media/{photo_id}.jpg'
 
 
+# возвращает путь вк фото
+def get_photo_url_mob(photo_id: str) -> str:
+    path = os.path.join(MEDIA_ROOT, f'{photo_id}.jpg')
+    if not os.path.exists(path):
+        download_file(photo_id, path)
+    return f'/media/{photo_id}.jpg'
