+++ standupuz_web/templates/Index_about.html
@@ -71,8 +71,8 @@
                         <a href="tel:+998333333333">+ 998 33 333 33 33</a>
                     </div>
                     <div class="header_right_contacts">
-                        <a href="#"><img src="../static/site/img/Instagram.svg" alt=""></a>
-                        <a href="#"><img src="../static/site/img/Telegram.svg" alt=""></a>
+                        <a href="https://instagram.com/standup.uz"><img src="../static/site/img/Instagram.svg" alt=""></a>
+                        <a href="https://t.me/StandUp_UZB"><img src="../static/site/img/Telegram.svg" alt=""></a>
                     </div>
                 </div>
             </div>
