+++ standupuz_web/templates/index_affiche.html
@@ -73,6 +73,7 @@
     <script>var cards = {{ cards|safe }};</script>
     <!-- <script src="../static/site/js/demo_cards_standupuz.js"></script> -->
     <script type="text/javascript" src="../static/site/js/main.js"></script>
+    <script type="text/javascript" src="../static/site/js/main_cards.js"></script>
    
 </body>
