+++ standupuz_web/templates/index_affiche_mob.html
@@ -3,43 +3,44 @@
 <head>
     <meta charset="UTF-8">
     <meta name="viewport" content="width=device-width, initial-scale=1.0">
-    <link rel="shortcut icon" href="../static/site/img/icon_black.png" type="image/x-icon">
-    <link rel="stylesheet" href="../static/site/css/basic_style.css">
-    <link rel="stylesheet" href="../static/site/css/sidenav_modal.css">
-    <link rel="stylesheet" href="../static/site/css/style.css">
-    <link rel="stylesheet" href="../static/site/css/media.css">
+    {% load static %}
+    <link rel="shortcut icon" href="{% static 'site/img/icon_black.png' %}" type="image/x-icon">
+    <link rel="stylesheet" href="{% static 'site/css/basic_style.css' %}">
+    <link rel="stylesheet" href="{% static 'site/css/sidenav_modal.css' %}">
+    <link rel="stylesheet" href="{% static 'site/css/style.css' %}">
+    <link rel="stylesheet" href="{% static 'site/css/media.css' %}">
     <title>StandUp - Афиша</title>
 </head>
 <body>
     
     <ul id="slide-out" class="sidenav">
-        <li><a class="sidenav-close" href="#!"><img src="../static/site/img/Vector.svg" alt=""></a></li>
-        <li><a class="sidenav-close" href="../events">Афиша</a></li>
-        <li><a class="sidenav-close" href="../about">О проекте</a></li>
+        <li><a class="sidenav-close" href="#!"><img src="{% static 'site/img/Vector.svg' %}" alt=""></a></li>
+        <li><a class="sidenav-close" href="{% url 'events' %}">Афиша</a></li>
+        <li><a class="sidenav-close" href="{% url 'about' %}">О проекте</a></li>
     </ul>
 
     <header class="header">
         <div class="conteiner">
             <div class="header_flex">
                 <div class="header_logo">
-                    <img class="logo" src="../static/site/img/WhiteSUz_wof_Png 2.svg" alt="Stadup">
+                    <img class="logo" src="{% static 'site/img/WhiteSUz_wof_Png 2.svg' %}" alt="Stadup">
                 </div>
                 <div class="header_menu">
                     <ul class="header_menu_ul">
-                        <li><a class="menu_activ" href="../events">Афиша</a></li>
-                        <li><a class="menu_notActiv" href="../about">О проекте</a></li>
+                        <li><a class="menu_activ" href="{% url 'events' %}">Афиша</a></li>
+                        <li><a class="menu_notActiv" href="{% url 'about' %}">О проекте</a></li>
                     </ul>
                 </div>
                 <div class="header_right">
                     <div class="header_right_burger">
-                        <a href="#" data-target="slide-out" class="sidenav-trigger"><i class="material-icons"><img src="../static/site/img/Menu.svg" alt="menu"></i></a>
+                        <a href="#" data-target="slide-out" class="sidenav-trigger"><i class="material-icons"><img src="{% static 'site/img/Menu.svg' %}" alt="menu"></i></a>
                     </div>
                     <div class="header_right_phone phone">
                         <a href="tel:+998333333333">+ 998 33 333 33 33</a>
                     </div>
                     <div class="header_right_contacts">
-                        <a href="#"><img src="../static/site/img/Instagram.svg" alt=""></a>
-                        <a href="#"><img src="../static/site/img/Telegram.svg" alt=""></a>
+                        <a href="https://instagram.com/standup.uz"><img src="{% static 'site/img/Instagram.svg' %}" alt=""></a>
+                        <a href="https://t.me/StandUp_UZB"><img src="{% static 'site/img/Telegram.svg' %}" alt=""></a>
                     </div>
                 </div>
             </div>
@@ -57,9 +58,11 @@
         </div>
     </div>
 
-    <script type="text/javascript" src="../static/site/js/materialize.min.js"></script>
-    <script src="../static/site/js/demo_cards_standupuz.js"></script>
-    <script type="text/javascript" src="../static/site/js/main.js"></script>
+    <script type="text/javascript" src="{% static 'site/js/materialize.min.js' %}"></script>
+    <script>var card = {{ card|safe }};</script>
+    <!-- <script src="../static/site/js/demo_cards_standupuz.js"></script> -->
+    <script type="text/javascript" src="{% static 'site/js/main.js' %}"></script>
+    <script type="text/javascript" src="{% static 'site/js/main_mob.js' %}"></script>
 
 </body>
 </html>
